Implemented vec_pop safely

+++ include/io.h
@@ -9,7 +9,7 @@ typedef struct {
     int closed;
 } File;
 
-long long get_filesize(const char *filename);
+int get_filesize(const char *filename);
 Result file_open(File* out, const char* path, const char* mode);
 Result file_close(File *in);
 Result read_to_string(const char *path, char** buffer);
