+++ src/io.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/stat.h>
@@ -6,12 +7,12 @@
 #include "logging.h"
 #include "result.h"
 
-long long get_filesize(const char *filename) {
+int get_filesize(const char *filename) {
     struct stat st;
 
     if(stat(filename, &st) == 0) {
         LOG_DEBUG("get_filesize: %s = %lld bytes", filename, (long long)st.st_size);
-        return (long long)st.st_size;
+        return (int)st.st_size;
     } else {
         LOG_WARN("get_filesize: failed to stat file %s", filename);
         return -1;
@@ -62,7 +63,7 @@ Result read_to_string(const char *path, char **out_buffer) {
         return r;
     }
 
-    long long size = get_filesize(path);
+    int size = get_filesize(path);
     if (size < 0) {
         LOG_ERROR("read_to_string error: file size was negative for '%s'", path);
         file_close(&file);
@@ -70,21 +71,21 @@ Result read_to_string(const char *path, char **out_buffer) {
     }
 
     LOG_DEBUG("read_to_string: allocating buffer of size %lld", size + 1);
-    char *buffer = malloc(size + 1);
+    char *buffer = malloc((size_t)size + 1);
     if (!buffer) {
         LOG_ERROR("read_to_string error: failed to allocate buffer for file '%s'", path);
         file_close(&file);
         return RESULT_ERR(ERR_MEMORY);
     }
 
-    size_t read = fread(buffer, 1, size, file.handle);
+    size_t read = fread(buffer, 1, (size_t)size, file.handle);
     if (read != (size_t)size) {
         if (feof(file.handle))
-            LOG_ERROR("read_to_string: unexpected EOF while reading %zu bytes from '%s'", size, path);
+            LOG_ERROR("read_to_string: unexpected EOF while reading %d bytes from '%s'", size, path);
         else if (ferror(file.handle))
-            LOG_ERROR("read_to_string: IO error while reading %zu bytes from '%s'", size, path);
+            LOG_ERROR("read_to_string: IO error while reading %d bytes from '%s'", size, path);
         else
-            LOG_ERROR("read_to_string: unknown read error (%zu/%zu bytes read) from '%s'", read, size, path);
+            LOG_ERROR("read_to_string: unknown read error (%zu/%d bytes read) from '%s'", read, size, path);
 
         free(buffer);
         file_close(&file);
@@ -94,7 +95,7 @@ Result read_to_string(const char *path, char **out_buffer) {
     buffer[size] = '\0';
     *out_buffer = buffer;
 
-    LOG_INFO("read_to_string: successfully read %lld bytes from '%s'", size, path);
+    LOG_INFO("read_to_string: successfully read %d bytes from '%s'", size, path);
     file_close(&file);
     return RESULT_OK();
 }
